[TS-335] Refactor: habit directory refactoring
