+++ src/components/Habit/Modal/StarPrizeModal/StarPrizeModal.style.ts
