+++ src/pages/HabitPage/HabitPage.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 
 import Gnb from "@/components/common/Gnb/Gnb";
+import HabitAdviceBanner from "@/components/Habit/HabitAdviceBanner/HabitAdviceBanner";
 import HabitGuideBanner from "@/components/Habit/HabitGuideBanner/HabitGuideBanner";
-import Advices from "@/components/Habit/Landing/Advices/Advices";
 import Calender from "@/components/Habit/Landing/Calendar/Calender";
 import Habits from "@/components/Habit/Landing/HabitList/HabitList";
 import Profile from "@/components/Habit/Profile/Profile";
@@ -43,7 +43,7 @@ const HabitPage = () => {
 				<div css={mainTopBoxStyle}>
 					<Profile />
 					<HabitGuideBanner />
-					<Advices />
+					<HabitAdviceBanner />
 				</div>
 				<Calender setSelectedDate={setSelectedDate} selectedDate={selectedDate} timeGap={timeGap} />
 				<Habits selectedDate={selectedDate} />
